Fix ロジック修正
