+++ src/features/user/components/LikedPictures.tsx
@@ -1,15 +1,36 @@
 import { memo } from 'react'
 
+import { Button } from '@/components/ui/button'
+import { Spinner } from '@/components/ui/spinner'
+
+import LoadingPictures from '@/features/pictures/components/LoadingPictures'
 import Pictures from '@/features/pictures/components/Pictures'
 
-import type { Picture } from '@/features/pictures/types'
+import { useInfiniteUserLikedPictures } from '../api'
 
 type LikedPicturesProps = {
-  likedPictures: Picture[]
+  userUid: string
 }
 
-const LikedPictures = memo(({ likedPictures }: LikedPicturesProps) => {
-  return <Pictures pictures={likedPictures} />
+const LikedPictures = memo(({ userUid }: LikedPicturesProps) => {
+  const useUserLikedPicturesQuery = useInfiniteUserLikedPictures({
+    userUid: userUid,
+  })
+  const userLikedPictures = useUserLikedPicturesQuery.data?.pages.flatMap((page) => page.data)
+
+  if (useUserLikedPicturesQuery.isLoading) return <LoadingPictures />
+  return (
+    <div className='mt-10'>
+      <Pictures pictures={userLikedPictures ?? []} />
+      {useUserLikedPicturesQuery.hasNextPage && (
+        <div className='flex items-center justify-center py-8'>
+          <Button onClick={() => useUserLikedPicturesQuery.fetchNextPage()} variant='outline'>
+            {useUserLikedPicturesQuery.isFetchingNextPage ? <Spinner /> : 'さらに読み込む'}
+          </Button>
+        </div>
+      )}
+    </div>
+  )
 })
 
 export default LikedPictures
